move yaml file operations to separate file
