+++ src/commands/apply.ts
@@ -3,7 +3,8 @@ import {BaseCommand} from '../lib/cmd'
 import color from '@oclif/color'
 import {apiCall} from '../lib/api'
 import {Config, getConfig} from '../lib/config'
-import {readYamlForAccountId, validateYaml, Yaml} from '../lib/yaml'
+import {validateYaml, Yaml} from '../lib/yaml'
+import {readYamlForAccountId} from '../lib/yaml-files'
 import {readYamlAtBranch} from '../lib/git'
 import {applySnowflake, findNotExistingEntities} from '../lib/spyglass'
 import {AppliedCommand} from '../lib/sql'
