+++ src/commands/verify.ts
@@ -2,7 +2,8 @@ import {Args, Flags, ux} from '@oclif/core'
 import {BaseCommand} from '../lib/cmd'
 import color from '@oclif/color'
 import {Config, getConfig} from '../lib/config'
-import {readYamlForAccountId, writeYamlForAccountId, Yaml} from '../lib/yaml'
+import {Yaml} from '../lib/yaml'
+import {readYamlForAccountId, writeYamlForAccountId} from '../lib/yaml-files'
 import {Issue, IssueDetail, ISSUE_HANDLERS} from '../lib/issues'
 import {printYamlDiff} from '../lib/print'
