+++ test/commands/import.test.ts
@@ -1,7 +1,8 @@
 import * as spyglass from '../../src/lib/spyglass'
 import * as yaml from '../../src/lib/yaml'
+import * as yamlFiles from '../../src/lib/yaml-files'
 import {expect, test} from '@oclif/test'
-import {readYamlFile} from '../../src/lib/yaml'
+import {readYamlFile} from '../../src/lib/yaml-files'
 
 let mockSpyglass = new spyglass.MockSpyglass()
 let mockYamlOutput: yaml.Yaml | null = null
@@ -31,7 +32,7 @@ describe('import', () => {
   .spyglass()
   .do(mockImportWithFile('./test/testdata/import-basic.yaml'))
   .stub(spyglass, 'newSpyglass', () => mockSpyglass)
-  .stub(yaml, 'writeYamlForAccountId', mockWriteYamlForAccountId as () => any)
+  .stub(yamlFiles, 'writeYamlForAccountId', mockWriteYamlForAccountId as () => any)
   .command(['import', 'account-123'])
   .exit(0)
   .it('runs import', ctx => {
